[W7][92] MEASURED: real GRPO drift produces ZERO staleness over 16 steps -- per-step requant is pure overhead there
Stage 1: their release stack run real (2xH100, their Qwen2.5-7B recipe,
16 steps, policy dumps per step; learned to reward 0.36). Stage 2: our
stack, drafter@0 vs every dump on the training prompt distribution,
64 prompts x 2 seeds: stale accept flat 3.93->3.98; fresh drafter@8/@16
-0.08 BELOW stale; detector gate never trips (correctly silent). Real
16-step drift maps to eps<=0.04 on the phase-89 knob -> the eps .25/.35
arms are the adversarial regime, disclosed as such. 2-GPU memory recipe:
FSDP offload + dynamic token-budget batching; expandable_segments is
incompatible with vLLM sleep-mode CuMem (measured). Their repo bug:
gitignore data/ swallowed verl/trainer/config/data (restored upstream).

+++ research/92_live_grpo/scripts/e1_staleness_curve.py
@@ -49,7 +49,8 @@ def main():
         spec = {"method": "draft_model", "model": DRAFT,
                 "num_speculative_tokens": K,
                 "draft_tensor_parallel_size": 1}
-    llm = LLM(model=TARGET, speculative_config=spec,
+    # veRL dumps are fp32; serve bf16 exactly like the veRL rollout does
+    llm = LLM(model=TARGET, speculative_config=spec, dtype="bfloat16",
               tensor_parallel_size=1, max_model_len=8192,
               gpu_memory_utilization=0.90, max_num_seqs=16,
               enable_prefix_caching=False, disable_log_stats=False,
